feat: finished Colors exercise. Started OrderManagement
Finished logic for Colors exercise.
Started the OrderManagement exercise, here the theory of the course showed that it is a good practice to use the Expert Pattern, I combined it with almost fully layer pattern, divided the folders with services, models , enums and the Application.java itself. Need to finished this exercise, it is pretty long